WIP: Add get, get?, all methods to Kemal::ParamParser
I appreciate this may be a controversial proposal, hence why I've
submitted it as WIP.

Taking inspiration from other frameworks, it would be handy if it was
possible to retrieve request parameters from any/all available sources
(url, query, json and body) without having to add
method/content-type checking boilerplate for every call.

This PR adds three methods, `get`, `get?` and `all`, which make it
much easier to retrieve values, as well as specifying a fallback in the
case of `get`. An example:

```crystal
get "/test/:name" do |env|
  # Returns value or nil
  name = env.params.get?("name").as(String)
  # => molovo

  # Returns value or default
  page = env.params.get("page", 1).as(Int32)
  # => 1

  data = env.params.all
  # => {
         "name" => "molovo",
         "v"    => 2
         "page" => 1
       }
end
```

There is obviously a slight compromise in terms of performance, but
given the number of request parameters in the average HTTP request, this
is likely to be minimal. Even so, if it is clearly documented, then those
looking for optimal performance can still continue to use the previous
methods, which haven't been changed.

I haven't written tests yet, since I wanted to see if this was something
the community was interested in first before putting too much time into
it. If the consensus is yes, then I'll get the tests written so that
this can be pushed through. Your thoughts, please!