Add decompress option also to other tests concerning gzip
The update concerns the other two tests concerning gzip. As of now, the response was
being automatically ungzipped if ContentEncoding was set to gzip, rendering the two
tests basically useless.
Adding decompress: false to both of them actually does not remove the ContentEncoding
header, so that we can ensure that the response is not actually gzipped.